Bug #85252: DataLayerCodeGenerator - při přesunu přepisuje existující repository

+++ Havit.Data.EntityFrameworkCore.CodeGenerator/Services/CodeWriter.cs
@@ -26,18 +26,16 @@ public class CodeWriter : ICodeWriter
 		// Abychom předešli problémům níže s tím, zda AlreadyExistsTheSameAsync má vrátit true nebo false, provedeme zde kontrolu, zda soubor existuje
 		// s a bez ohledu na case-sensitivitu a případně soubor přejmenujeme na správný název (case sensitivity).
 		// Názvy složek však nejsou řešeny.
-		bool exists = File.Exists(filename);
+		bool existsCaseInsensitive = FileExistsCaseInsensitive(filename);
 
-		if (exists && RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+		bool existsCaseSensitive = existsCaseInsensitive && FileExistsCaseSensitive(filename); // exists: nemuže existovat case sensitive, pokud vůbec neexistuje
+		if (existsCaseInsensitive && !existsCaseSensitive)
 		{
-			// mohou se lišit jen ve windows, jinde jsou vždy stejné
-			bool existsCaseSensitive = FileExistsCaseSensitive(filename); // exists: nemuže existovat case sensitive, pokud vůbec neexistuje
-			if (!existsCaseSensitive)
-			{
-				// přejmenuje soubor na správný název (bez ohledu na název složky)
-				System.IO.File.Move(filename, filename);
-			}
+			// přejmenuje soubor na správný název (bez ohledu na název složky)
+			System.IO.File.Move(filename, filename);
+			existsCaseSensitive = true;
 		}
+		// nyní by existsCaseInsensitive mělo mít stejnou hodnotu, jako existsCaseSensitive
 
 		if (!(await AlreadyExistsTheSameAsync(filename, content, cancellationToken)) || !HasByteOrderMask(filename))
 		{
@@ -47,7 +45,7 @@ public class CodeWriter : ICodeWriter
 				Directory.CreateDirectory(directory);
 			}
 
-			if ((overwriteBahavior == OverwriteBahavior.OverwriteWhenFileAlreadyExists) || !exists)
+			if ((overwriteBahavior == OverwriteBahavior.OverwriteWhenFileAlreadyExists) || !existsCaseInsensitive)
 			{
 				await File.WriteAllTextAsync(filename, content, Encoding.UTF8, cancellationToken);
 			}
@@ -87,9 +85,18 @@ public class CodeWriter : ICodeWriter
 	/// <summary>
 	/// Vrací true, pokud existuje soubor se stejným názvem case-sensitive. Case sensitivita složky se neřeší.
 	/// </summary>
+	public static bool FileExistsCaseInsensitive(string fullPath)
+	{
+		string filename = Path.GetFileName(fullPath);
+		return new DirectoryInfo(Path.GetDirectoryName(fullPath)).EnumerateFiles(filename).Any(file => String.Equals(file.Name, filename, StringComparison.CurrentCultureIgnoreCase));
+	}
+
+	/// <summary>
+	/// Vrací true, pokud existuje soubor se stejným názvem case-sensitive. Case sensitivita složky se neřeší.
+	/// </summary>
 	public static bool FileExistsCaseSensitive(string fullPath)
 	{
 		string filename = Path.GetFileName(fullPath);
-		return new DirectoryInfo(Path.GetDirectoryName(fullPath)).EnumerateFiles(filename).Any(file => file.Name == filename);
+		return new DirectoryInfo(Path.GetDirectoryName(fullPath)).EnumerateFiles(filename).Any(file => String.Equals(file.Name, filename, StringComparison.CurrentCulture));
 	}
 }
